Move form database migration code to its own class
